Rework to separate the single /etc/xap-livebox.ini file into separate configuration files all beneath the /etc/xap.d/ directory.
The second part of the naming is now based upon the HOSTNAME of the livebox.
By default this is "livebox" so an update to this release will have no affect.
dbzoo.livebox.STUFF

However if you have multiple livebox's on your network you should name each
differently so they are unique.
dbzoo.livebox2.STUFF

This change prepares the code base for porting to another hardware platform.

Configuration files will automigrate upon upgrading to this release: 311.1


git-svn-id: https://livebox-hah.googlecode.com/svn/trunk@511 64644fc6-01da-11df-bca2-d32778fd5087